requirements.txt
ipykernel
opencv-python
torch
numpy
pandas
seaborn 
matplotlib
scikit-learn
tqdm
fastapi
pydantic
python-dotenv
google-generativeai
uvicorn
opencv-python
pillow
torchvision
python-multipart
uvicorn[standard]
fastapi